Iteration / Bug fixes
-Polygon now initializes selected to false
-Changed rotation math to use doubles
-ObjectEditor assigns first poly automatically for transformations if
none has been selected yet

+++ ObjectEditor.h
@@ -34,7 +34,7 @@ public:
     
     static void TranslatePolygon(Vector2i translationVector, bool drawScene);
     static void ScalePolygon(float scaleX, float scaleY, bool drawScene);
-    static void RotatePolygon(float degrees, bool drawScene);
+    static void RotatePolygon(double degrees, bool drawScene);
 
     static deque<Polygon> GetPolygons();
